mapping-service/pyproject.toml
[project]
name = "mining-microservice"
version = "0.1.0"
description = "LLM-based microservice for document mining"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi",
    "uvicorn",
    "pydantic",
    "requests",
    "python-dotenv",
    "mangum"
]